Relocate ServiceMutatorsGenerator to generators package
