+++ internal/generators/templates.go
@@ -17,4 +17,7 @@ var (
 
 	//go:embed templates/partial_service_method_impl_to_private.go.tmpl
 	templateServiceMethodImplToPrivate string
+
+	//go:embed templates/partial_service_mutators.go.tmpl
+	templateServiceMutators string
 )
