--- internal/service_mutators_generator.go
@@ -1,11 +1,0 @@
-package internal
-
-type ServiceMutatorGenerator struct {
-	MessageName string
-	Fields      []MutatorField
-}
-
-type MutatorField struct {
-	FieldName string
-	FieldType string
-}
