+++ internal/templates.go
@@ -1,17 +1,11 @@
 package internal
 
 import (
-	_ "embed"
 	"fmt"
 	"io"
 	"text/template"
 
 	"github.com/dane/protoc-gen-go-svc/internal/generators"
-)
-
-var (
-	//go:embed templates/partial_service_mutators.go.tmpl
-	templateServiceMutators string
 )
 
 func execute(name string, templateStr string, w io.Writer, params interface{}) error {
