--- internal/templates/partial_service_mutators.go.tmpl
@@ -1,17 +1,0 @@
-{{ range $message := . }}
-	type {{ .MessageName }}Mutator func(*privatepb.{{ .MessageName }})
-
-	func Apply{{ .MessageName }}Mutators(in *privatepb.{{ .MessageName }}, mutators []{{ .MessageName }}Mutator) {
-		for _, mutator := range mutators {
-			mutator(in)
-		}
-	}
-
-	{{ range $field := .Fields -}}
-		func Set{{ $message.MessageName }}_{{ $field.FieldName }}(value {{ $field.FieldType }}) {{ $message.MessageName }}Mutator {
-			return func(in *privatepb.{{ $message.MessageName }}) {
-				in.{{ $field.FieldName }} = value
-			}
-		}
-	{{ end -}}
-{{ end }}
